# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EasyAudit-Next API",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/review-cases": {
      "get": {"operationId": "listReviewCases", "tags": ["review-planning"]},
      "post": {"operationId": "createReviewCase", "tags": ["review-planning"], "responses": ["201", "409", "422"]}
    },
...
    "/api/v1/management/review-cases/export": {
      "get": {
        "operationId": "exportManagedReviewCases",
        "tags": ["management"],
        "responses": ["200", "422"]
      }
    }
  }
}